updating context
