+++ client/src/Components/RightPane.js
@@ -1,5 +1,6 @@
 import React, { useEffect, useState, useContext } from 'react';
-import ChartContex from './ChartContext';
+import { CheckinContext } from './CheckinContext';
+import { UserContext } from './UserContext';
 import { Grid, Paper, Typography } from '@material-ui/core';
 import Reminders from './Reminders';
 import LineChart from './Chart';
@@ -7,7 +8,38 @@ import API from '../Utils/clientauth';
 import moment from 'moment';
 
 const RightPane = props => {
-  const { chartdata, dates, chartRendered } = useContext(ChartContext);
+  const { user } = useContext(UserContext);
+  const {
+    chartdata,
+    setChartData,
+    dates,
+    setDates,
+    chartRendered,
+    setChartRendered
+  } = useContext(CheckinContext);
+
+  const updateUserData = () => {
+    if (user) {
+      let checkinPoints = [];
+      let dates = [];
+      API.getUserData(user._id)
+        .then(res => {
+          let points = res.data;
+          for (let i = 0; i < points.length; i++) {
+            checkinPoints.push(points[i].totalPoints);
+            dates.push(moment(points[i].date).format('MMM D'));
+          }
+          setChartData(checkinPoints);
+          setDates(dates);
+          setChartRendered(true);
+        })
+        .catch(err => {
+          console.log(err);
+        });
+    } else {
+      // need else code
+    }
+  };
 
   return (
     <Grid container md={7} sm={12} xs={12}>
@@ -15,10 +47,7 @@ const RightPane = props => {
         <Paper elevation={3} style={this.props.style}>
           <Typography style={this.props.header}>Daily Dashboard</Typography>
           {chartRendered === false ? null : (
-            <LineChart
-              chartdata={chartdata}
-              dates={dates}
-            />
+            <LineChart chartdata={chartdata} dates={dates} />
           )}
         </Paper>
       </Grid>
